refactor hydrophones sim side
