+++ Assets/Scripts/PingerBearingVisualisation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5df3b156aab9a98439920a7c4d803ae4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
